refactor(service-log): fold ui/ into presentation/ui
