+++ src/module/car/presentation/ui/sections/overview/overview.tsx
@@ -1,7 +1,7 @@
 import { OwnershipsSection } from '@/car/ownership/presentation/ui/sections/ownerships/ownerships';
 import { DetailsSection } from '@/car/presentation/ui/sections/details/details';
 import { CarCostsSection } from '@/car/service-log/presentation/ui/sections/car-costs/car-costs';
-import { ServiceLogsSection } from '@/car/service-log/ui/sections/service-logs/service-logs';
+import { ServiceLogsSection } from '@/car/service-log/presentation/ui/sections/service-logs/service-logs';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 
 import { DeleteSection } from '../delete/delete';
